third_party/blink/web_tests/editing/selection/caret-after-removing-node-expected.txt

PASS getSelection().type is "Caret"
PASS getSelection().focusNode is sample
PASS getSelection().focusOffset is 0
PASS successfullyParsed is true

TEST COMPLETE
